🐛 Fix the local authentication
The session cookie was not properly set by the server due to the CORS policy. So every requests after the authentication was failing, as the server wouldn't recognize the user.
Easy fix for the local environment: set the exact same domain name for the API calls than the Vite server. Even if localhost points the the same directory than 127.0.0.1, the browser don't consider them as equivalent, as the URL is different.
The Access-Control-Allow-Credentials header was also needed, server-side and client-side (the withCredentials parameter in Axios).