♻️ Move the loader logic of the organization creation form from Redux to a local state
The main goal of all that branch is to handle the current organization in the same reducer than the current user (to simplify the logout procedure). So to avoid confusion, the organization creation form needs to be managed as locally as possible. It's also a good practice in Redux to avoid storing form state in it.