🩹 Fix an error in console when creating a user
In the user/addUser thunk, the full Axios response was returned to the reducer, instead of the data property. It was not a problem for the app, because these data are not used by the reducer, but an error was thrown due to some non-serializable content in the Axios response being sent to Redux.