🚑 Fix the broken profile edit form
This form is used in two scenarios: in the sign up process, or when the user update its profile. But the organization name was retrieved from React Router in every case, which was leading to a fatal error.
Now this data is only fetched when signing up, just before the call of the API request.