💄 Add a loader while fetching the invitation in the sign-up form
Thanks to this, the user has a clear indication that something is being prepared in the background, and won't be able to fill in the form while it's not ready.
It also prevents a little bug with MUI form fields, where the field label wouldn't rise up if its value is set programmatically while it's displayed. That was occuring with the email field, which is automatically filled in after the invitation fetch.