🥅 🚸 Add some try catch blocks around the profile form sendings
Server validation errors will now be displayed below each corresponding fields. In a non-hacking scenario, only the email (when already taken), the old password (when incorrect) and the profile picture (when the file is not an image or when the dimensions are too small) are concerned.
Also, the global form error zone is now displayed just above the submit button, which is way more ergonomic.