🚸 Remove the redirection to the login page when a 401 error occurs
It was not working very well: if the error occurred when changing page, the 401 error page was appearing briefly before finally displaying the homepage and login form. This flickering was not a really pleasing experience.
When the error occurred after a request which did not involved a navigation, like loading the comments of a post for example, the homepage was not even shown, the user ended on the 401 error page directly.
Removing that redirection to the homepage makes this error handling consistent... and avoids any flickering.