🚸 Hide the title of the invitation page during the loading of the invitation
Since the previous commit, the title includes the name of the organization, which can only be known after the invitation has been fetched from the server. Due to this, the title was switching from the version of the admin sign up page to the one of the invitation form right after the loading was over.
The title is now hidden during the loading to avoid this flashing.