🐛 Fix a bug with the persistence of the modifications of the user profile
After using the form to update the user's profile and reloading the page (or just closing the tab or the browser), the modifications were lost on the client side. The current user is stored in localStorage to persist their data between page reloads, and this entry was simply not updated after submitting the form.
Using localStorage for this purpose is far from ideal, and this behavior will be changed in future commits. However, this quick fix eliminates that annoying bug!