🔒️ Removed the user data from localStorage
These data was stored in localStorage to persist the user between page reloads and after closing the tab or the browser, as long as the session was still active server-side. However, this system was only temporary due to a serious security issue. In fact, localStorage data can easily be modified by the end user, and nothing had been implemented to check the validity of the data.
Now, the current user is simply fetched from the server before doing anything else, preventing the security issues mentioned above, and also ensuring the user data is up to date if it was modified on another device or browser.