♻️ Set the default currentPage value of the posts pagination to 0 instead of 1
Previously, this currentPage value was set to 1 while the posts were loading, then incremented to 2 once loaded. This behavior was confusing, as it implied the property was more like a "nextPage" indicator.
Now, currentPage starts at 0 during loading, incrementing to 1 once the first page of posts is loaded, which aligns better with its name and expected role.