🐛 Fix the incorrect display of the "no more posts" message when an activity feed doesn't contain any post
In an activity feed, a placeholder message indicates that all posts have been loaded when a profile or an organization contains some. However, it was incorrectly displayed on empty feeds, appearing right after the "no posts at all" message, following the first server request.
Now, this message only appears when it's relevant.