♻️ Simplify the code of the loader and the placeholder messages of the posts infinite scroll
Some conditions were repeated and the overall logic for displaying these elements was too dispersed. The code is now more concise and easier to read.
Additionnally, the "no posts" message is now centered, as it's now placed within the same flex Box element as the loader and the other placeholder.