♻️ Rename the organization name field and the related state in the OrganizationForm component
The pervious name, organizationName, was descriptive but differed from the server field. Renaming it just "name" simplifies the code and aligns with the API field, making it easier to work with server data.
It's also necessary for the next commits, where server errors related to the name field will be displayed beneath it, using the error and helperText attributes of MUI's TextField component.