🔥 Simplify the code used to navigate to the sign-up page in the OrganizationForm component
The useEffect was necessarry in an earlier version of the component, where the organization form treatments were handled in Redux. Since the API request is now called directly in this component, the navigate can be launched right after receiving the response.
This change simplifies the code, removing the need for the useEffect and the name state.