💄 Increase the width of the organization form input
This input was too narrow and not limited in width. This caused it to expand abruptly when the server error message (in case of name conflict) was added beneath it, which was visually disruptive.
Now, it takes the full width of its container to avoid this undesired effect.
Additionnally, the form itself and its titles have been reajusted to align with the input size, resulting in a cleaner appearance.