🚨 Fix an unwanted ESLint warning with the indentation of switch/case blocks
The indent rule (set in the 9dd956d commit) required an additionnal setting for the switch/case blocks. Without it, a warning was triggered if the case blocks weren't on the same column than their parent switch, which was confusing as the best practice is to indent them by 1 level.
This is the purpose of the "SwitchCase" setting by the way: "1" means "1 indent", regardless of the indentation size.