🚸 Change error pages logic to display them without changing URL
When an error occurs, the app displays an error page to the user (among 401, 403, 404 and 500). But until now, the user was redirected to one of these pages, which rewrote the URL. However, users should always be able to see the original URL that caused the error, so these redirects didn't provide an ideal experience.
Now, a new ErrorPageHandler component wraps the Router, and can be controlled through Redux to display either an error page or the intended component. All error pages of the app are now displayed through this system, so the URL is never rewritten.