🐛 Fix the missing back button on router 404 errors
Since the previous branch, all error pages are displayed through the ErrorPageHandler component. However, one part of the app was overlooked: the router 404 error, triggered by the wilcard path ("/*"). So when a URL that didn't match any configured route was requested, the 404 error page was displayed correctly but the back button from the ButtonHome component was missing.
To resolve this, the cleanest solution was to handle this error with the ErrorPageHandler, which required creating a new component, NotFoundRoute, to integrate with React Router.