🐛 Fix rare error in users/logout thunk
The thunkApi.rejectWithValue method call wasn't working because the thunkApi argument was misplaced in the createAsyncThunk callback. In fact, the first argument, usually the payload, had been omitted, meaning thunkApi was never passed to the callback.
Now, thunkApi is correctly set as the second argument, with just an underscore as the first argument, to indicate no payload is expected here. This was likely the root cause of the mistake by the way.