♻️ Make error code handling more generic in thunks
Previously, some thunks returned hardcoded error codes in the thunkApi.rejectWithValue method call. In some cases, this duplicated the code already detected in the preceding condition; in others, the error code was always forced to 500, regardless of the original one.
Now, to avoid redundancy and mostly to ensure accurate error codes, they are returned directly as they are.
Note: even though the ProfileForm component is not a thunk, it was included in the commit for the exact same reason as the thunks.