🧑‍💻 Rename several payload properties and variables of feed thunks to avoid confusion
In the feed/createReaction and feed/updateReaction thunks, the payload included a `reaction` property. Additionally, a `newReaction` variable was created with the server's feedback data. While both names implied they contained a complete reaction model, only `newReaction` did; `reaction` was merely a string representing the type of the new reaction.
Now, to eliminate confusion, these have been renamed to better reflect their content:
- `reaction` is now `type` (matching the server field by the way)
- `newReaction` is now simply `reaction` (dropping the `new` prefix, as it's no longer necessary).

Following the same principle, the `reaction` parameter of the `handleReaction` function in the `ReactionButton` component has been renamed to `type` as well, as the function had to be modified to apply the changes described above.