♻️ Move members list logic in admin pages from Redux to component level
Redux was overused in this case. While this library is designed for managing global or semi-global states, the members list is relevant only to two specific components which are visible exclusively:
- to administrators (a limited set of users)
- within the admin pages (a specific section of the app).

Now, the API requests and related logic are handled directly within the `AdminMembers` and `MemberCard` components, simplifying the code and reducing unnecessary global state management.