🔥 Remove unnecessary React Hook Form usage in `MemberCard` component
Although the `MemberCard` component can send HTTP requests via the "Block/Unblock" button, it is not a form. The use of React Hook Form, along with a `<form>` element and a `submit` type button, introduced unnecessary complexity and potential confusion.
Now, the HTTP request is triggered using a simple `onClick` prop on the `Button` element, making the implementation both simpler and more appropriate.